+++ packages/monaco/src/libs/theme/editorTheme.ts
@@ -1,0 +1,7 @@
+export enum EditorTheme {
+  // We may eventually want to change this to something like 'interface'
+  // instead of 'dark' since it will indicate that the theme should match
+  // the current LE color mode
+  dark = 'dark',
+  highContrast = 'hc-black',
+}