+++ packages/monaco/src/libs/theme/types.ts
@@ -1,0 +1,4 @@
+export type MonacoThemeColors = {
+  syntax: Record<string, string>;
+  ui: Record<string, string>;
+};