+++ packages/monaco/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.1.0 (2021-07-28)
 
-
 ### Features
 
-* **monaco:** add monaco package ([14600b9](http://github.com///commit/14600b90ec8930624ecbe6df750f4cbcb587b0e3))
+- **monaco:** add monaco package ([14600b9](http://github.com///commit/14600b90ec8930624ecbe6df750f4cbcb587b0e3))